Refactor to follow new pattern
