+++ bittensor/commands/__init__.py
@@ -17,15 +17,60 @@
 
 from munch import Munch, munchify
 
-
-from .stake import ( 
-    StakeCommand,  
-    StakeShow,
-    SetChildCommand, 
-    SetChildrenCommand,     
-    GetChildrenCommand,   
+defaults: Munch = munchify(
+    {
+        "netuid": 1,
+        "subtensor": {"network": "finney", "chain_endpoint": None, "_mock": False},
+        "pow_register": {
+            "num_processes": None,
+            "update_interval": 50000,
+            "output_in_place": True,
+            "verbose": False,
+            "cuda": {"dev_id": [0], "use_cuda": False, "tpb": 256},
+        },
+        "axon": {
+            "port": 8091,
+            "ip": "[::]",
+            "external_port": None,
+            "external_ip": None,
+            "max_workers": 10,
+            "maximum_concurrent_rpcs": 400,
+        },
+        "priority": {"max_workers": 5, "maxsize": 10},
+        "prometheus": {"port": 7091, "level": "INFO"},
+        "wallet": {
+            "name": "default",
+            "hotkey": "default",
+            "path": "~/.bittensor/wallets/",
+        },
+        "dataset": {
+            "batch_size": 10,
+            "block_size": 20,
+            "num_workers": 0,
+            "dataset_names": "default",
+            "data_dir": "~/.bittensor/data/",
+            "save_dataset": False,
+            "max_datasets": 3,
+            "num_batches": 100,
+        },
+        "logging": {
+            "debug": False,
+            "trace": False,
+            "record_log": False,
+            "logging_dir": "~/.bittensor/miners",
+        },
+    }
 )
-from .unstake import UnStakeCommand, RevokeChildCommand, RevokeChildrenCommand  
+
+
+from .stake.add import StakeCommand
+from .stake.get_children_info import GetChildrenCommand
+from .stake.set_child import SetChildCommand
+from .stake.set_children import SetChildrenCommand
+from .stake.show import StakeShow
+from .stake.revoke_child import RevokeChildCommand
+from .stake.revoke_children import RevokeChildrenCommand
+from .unstake import UnStakeCommand 
 from .overview import OverviewCommand     
 from .register import (
     PowRegisterCommand,     
@@ -82,52 +127,7 @@ from .root import (
     RootSetBoostCommand,     
     RootSetSlashCommand,     
 )
-from .identity import GetIdentityCommand, SetIdentityCommand     
-
-defaults: Munch = munchify(
-    {
-        "netuid": 1,
-        "subtensor": {"network": "finney", "chain_endpoint": None, "_mock": False},
-        "pow_register": {
-            "num_processes": None,
-            "update_interval": 50000,
-            "output_in_place": True,
-            "verbose": False,
-            "cuda": {"dev_id": [0], "use_cuda": False, "tpb": 256},
-        },
-        "axon": {
-            "port": 8091,
-            "ip": "[::]",
-            "external_port": None,
-            "external_ip": None,
-            "max_workers": 10,
-            "maximum_concurrent_rpcs": 400,
-        },
-        "priority": {"max_workers": 5, "maxsize": 10},
-        "prometheus": {"port": 7091, "level": "INFO"},
-        "wallet": {
-            "name": "default",
-            "hotkey": "default",
-            "path": "~/.bittensor/wallets/",
-        },
-        "dataset": {
-            "batch_size": 10,
-            "block_size": 20,
-            "num_workers": 0,
-            "dataset_names": "default",
-            "data_dir": "~/.bittensor/data/",
-            "save_dataset": False,
-            "max_datasets": 3,
-            "num_batches": 100,
-        },
-        "logging": {
-            "debug": False,
-            "trace": False,
-            "record_log": False,
-            "logging_dir": "~/.bittensor/miners",
-        },
-    }
-)
+from .identity import GetIdentityCommand, SetIdentityCommand   
 
 __all__ = [
     "StakeCommand",
