+++ bittensor/extrinsics/root.py
@@ -177,7 +177,7 @@ def set_root_weights_extrinsic(
         )
     ):
         try:
-            weight_uids, weight_vals = weight_utils.convert_values_and_ids_for_emit(
+            weight_uids, weight_vals = weight_utils.convert_weights_and_uids_for_emit(
                 netuids, weights
             )
             success, error_message = subtensor._do_set_root_weights(
