+++ bittensor/extrinsics/set_weights.py
@@ -64,7 +64,7 @@ def set_weights_extrinsic(
             If ``true``, the call waits for confirmation from the user before proceeding.
     Returns:
         success (bool):
-            Flag is ``true`` if extrinsic was finalized or uncluded in the block. If we did not wait for finalization / inclusion, the response is ``true``.
+            Flag is ``true`` if extrinsic was finalized or included in the block. If we did not wait for finalization / inclusion, the response is ``true``.
     """
     # First convert types.
     if use_torch():
@@ -79,7 +79,7 @@ def set_weights_extrinsic(
             weights = np.array(weights, dtype=np.float32)
 
     # Reformat and normalize.
-    weight_uids, weight_vals = weight_utils.convert_values_and_ids_for_emit(
+    weight_uids, weight_vals = weight_utils.convert_weights_and_uids_for_emit(
         uids, weights
     )
