+++ bittensor/extrinsics/staking.py
@@ -19,11 +19,8 @@
 import bittensor
 from rich.prompt import Confirm
 from time import sleep
-import numpy as np
 from typing import List, Union, Optional, Tuple
-from numpy.typing import NDArray
 
-from bittensor.utils import weight_utils
 from bittensor.utils.balance import Balance
 
 
@@ -527,222 +524,3 @@ def __do_add_stake_single(
     )
 
     return success
-
-
-def do_set_child_singular_extrinsic(
-    subtensor: "bittensor.subtensor",
-    wallet: "bittensor.wallet",
-    hotkey: str,
-    child: str,
-    netuid: int,
-    proportion: float,
-    wait_for_inclusion: bool = True,
-    wait_for_finalization: bool = False,
-    prompt: bool = False,
-) -> bool:
-    r"""
-    Sets child hotkey with a proportion assigned from the parent.
-
-    Args:
-        wallet (bittensor.wallet):
-            Bittensor wallet object.
-        hotkey (str):
-            Parent hotkey.
-        child (str):
-            Child hotkey.
-        netuid (int):
-            Unique identifier of for the subnet.
-        proportion (float):
-            Proportion assigned to child hotkey.
-        wait_for_inclusion (bool):
-            If set, waits for the extrinsic to enter a block before returning ``true``, or returns ``false`` if the extrinsic fails to enter the block within the timeout.
-        wait_for_finalization (bool):
-            If set, waits for the extrinsic to be finalized on the chain before returning ``true``, or returns ``false`` if the extrinsic fails to be finalized within the timeout.
-        prompt (bool):
-            If ``true``, the call waits for confirmation from the user before proceeding.
-    Returns:
-        success (bool):
-            Flag is ``true`` if extrinsic was finalized or included in the block. If we did not wait for finalization / inclusion, the response is ``true``.
-    Raises:
-        bittensor.errors.ChildHotkeyError:
-            If the extrinsic fails to be finalized or included in the block.
-        bittensor.errors.NotRegisteredError:
-            If the hotkey is not registered in any subnets.
-
-    """
-    # Ask before moving on.
-    if prompt:
-        if not Confirm.ask(
-            "Do you want to add child hotkey:\n[bold white]  child: {}\n  proportion: {}[/bold white ]?".format(
-                child, proportion
-            )
-        ):
-            return False
-
-    with bittensor.__console__.status(
-        ":satellite: Setting child hotkey on [white]{}[/white] ...".format(
-            subtensor.network
-        )
-    ):
-        try:
-            # prepare values for emmit
-            proportion = np.array([proportion], dtype=np.float32)
-            netuids = np.full(proportion.shape, netuid, dtype=np.int64)
-
-            uid_val, proportion_val = weight_utils.convert_values_and_ids_for_emit(
-                netuids, proportion
-            )
-
-            success, error_message = subtensor._do_set_child_singular(
-                wallet=wallet,
-                hotkey=hotkey,
-                child=child,
-                netuid=netuid,
-                proportion=proportion_val[0],
-                wait_for_inclusion=wait_for_inclusion,
-                wait_for_finalization=wait_for_finalization,
-            )
-
-            bittensor.__console__.print(success, error_message)
-
-            if not wait_for_finalization and not wait_for_inclusion:
-                return True
-
-            if success is True:
-                bittensor.__console__.print(
-                    ":white_heavy_check_mark: [green]Finalized[/green]"
-                )
-                bittensor.logging.success(
-                    prefix="Set child hotkey",
-                    suffix="<green>Finalized: </green>" + str(success),
-                )
-                return True
-            else:
-                bittensor.__console__.print(
-                    f":cross_mark: [red]Failed[/red]: {error_message}"
-                )
-                bittensor.logging.warning(
-                    prefix="Set child hotkey",
-                    suffix="<red>Failed: </red>" + str(error_message),
-                )
-                return False
-
-        except Exception as e:
-            bittensor.__console__.print(
-                ":cross_mark: [red]Failed[/red]: error:{}".format(e)
-            )
-            bittensor.logging.warning(
-                prefix="Set child hotkey", suffix="<red>Failed: </red>" + str(e)
-            )
-            return False
-
-
-def do_set_children_multiple_extrinsic(
-    subtensor: "bittensor.subtensor",
-    wallet: "bittensor.wallet",
-    hotkey: str,
-    children: Union[NDArray[str], list],
-    netuid: int,
-    proportions: Union[NDArray[np.float32], list],
-    wait_for_inclusion: bool = True,
-    wait_for_finalization: bool = False,
-    prompt: bool = False,
-) -> bool:
-    r"""
-    Sets children hotkeys with a proportion assigned from the parent.
-
-    Args:
-        wallet (bittensor.wallet):
-            Bittensor wallet object.
-        hotkey (str):
-            Parent hotkey.
-        children (np.ndarray):
-            Children hotkeys.
-        netuid (int):
-            Unique identifier of for the subnet.
-        proportions (np.ndarray):
-            Proportions assigned to children hotkeys.
-        wait_for_inclusion (bool):
-            If set, waits for the extrinsic to enter a block before returning ``true``, or returns ``false`` if the extrinsic fails to enter the block within the timeout.
-        wait_for_finalization (bool):
-            If set, waits for the extrinsic to be finalized on the chain before returning ``true``, or returns ``false`` if the extrinsic fails to be finalized within the timeout.
-        prompt (bool):
-            If ``true``, the call waits for confirmation from the user before proceeding.
-    Returns:
-        success (bool):
-            Flag is ``true`` if extrinsic was finalized or included in the block. If we did not wait for finalization / inclusion, the response is ``true``.
-    Raises:
-        bittensor.errors.ChildHotkeyError:
-            If the extrinsic fails to be finalized or included in the block.
-        bittensor.errors.NotRegisteredError:
-            If the hotkey is not registered in any subnets.
-
-    """
-    # Ask before moving on.
-    if prompt:
-        if not Confirm.ask(
-            "Do you want to add children hotkeys:\n[bold white]  children: {}\n  proportions: {}[/bold white ]?".format(
-                children, proportions
-            )
-        ):
-            return False
-
-    with bittensor.__console__.status(
-        ":satellite: Setting children hotkeys on [white]{}[/white] ...".format(
-            subtensor.network
-        )
-    ):
-        try:
-            # prepare values for emmit
-            if isinstance(proportions, np.ndarray):
-                uids = np.full(proportions.shape, netuid, dtype=np.int64)
-            else:
-                uids = [netuid] * len(proportions)
-
-            uid_val, proportions_val = weight_utils.convert_values_and_ids_for_emit(
-                uids, proportions
-            )
-
-            children_with_proportions = list(zip(children, proportions_val))
-
-            success, error_message = subtensor._do_set_children_multiple(
-                wallet=wallet,
-                hotkey=hotkey,
-                children_with_proportions=children_with_proportions,
-                netuid=netuid,
-                wait_for_inclusion=wait_for_inclusion,
-                wait_for_finalization=wait_for_finalization,
-            )
-
-            bittensor.__console__.print(success, error_message)
-
-            if not wait_for_finalization and not wait_for_inclusion:
-                return True
-
-            if success is True:
-                bittensor.__console__.print(
-                    ":white_heavy_check_mark: [green]Finalized[/green]"
-                )
-                bittensor.logging.success(
-                    prefix="Set children hotkeys",
-                    suffix="<green>Finalized: </green>" + str(success),
-                )
-                return True
-            else:
-                bittensor.__console__.print(
-                    f":cross_mark: [red]Failed[/red]: {error_message}"
-                )
-                bittensor.logging.warning(
-                    prefix="Set children hotkeys",
-                    suffix="<red>Failed: </red>" + str(error_message),
-                )
-                return False
-
-        except Exception as e:
-            bittensor.__console__.print(
-                ":cross_mark: [red]Failed[/red]: error:{}".format(e)
-            )
-            bittensor.logging.warning(
-                prefix="Set children hotkeys", suffix="<red>Failed: </red>" + str(e)
-            )
-            return False
