+++ tests/e2e_tests/subcommands/stake/test_stake_add_remove.py
@@ -35,7 +35,7 @@ def test_stake_add(local_chain):
         == 0
     )
 
-    exec_command(RegisterCommand, ["s", "register", "--neduid", "1"])
+    exec_command(RegisterCommand, ["s", "register", "--netuid", "1"])
 
     assert (
         local_chain.query(
