+++ tests/unit_tests/utils/test_utils.py
@@ -27,33 +27,33 @@ import pytest
 def test_convert_weight_and_uids():
     uids = np.arange(10)
     weights = np.random.rand(10)
-    weight_utils.convert_values_and_ids_for_emit(uids, weights)
+    weight_utils.convert_weights_and_uids_for_emit(uids, weights)
 
     # min weight < 0
     weights[5] = -1
     with pytest.raises(ValueError) as pytest_wrapped_e:
-        weight_utils.convert_values_and_ids_for_emit(uids, weights)
+        weight_utils.convert_weights_and_uids_for_emit(uids, weights)
 
     # min uid < 0
     weights[5] = 0
     uids[3] = -1
     with pytest.raises(ValueError) as pytest_wrapped_e:
-        weight_utils.convert_values_and_ids_for_emit(uids, weights)
+        weight_utils.convert_weights_and_uids_for_emit(uids, weights)
 
     # len(uids) != len(weights)
     uids[3] = 3
     with pytest.raises(ValueError) as pytest_wrapped_e:
-        weight_utils.convert_values_and_ids_for_emit(uids, weights[1:])
+        weight_utils.convert_weights_and_uids_for_emit(uids, weights[1:])
 
     # sum(weights) == 0
     weights = np.zeros(10)
-    weight_utils.convert_values_and_ids_for_emit(uids, weights)
+    weight_utils.convert_weights_and_uids_for_emit(uids, weights)
 
     # test for overflow and underflow
     for _ in range(5):
         uids = np.arange(10)
         weights = np.random.rand(10)
-        weight_utils.convert_values_and_ids_for_emit(uids, weights)
+        weight_utils.convert_weights_and_uids_for_emit(uids, weights)
 
 
 def test_normalize_with_max_weight():
